broker updated

+++ src/services/broker.service.js
@@ -1,7 +1,65 @@
+const fs = require("fs");
+
+const path = require("path");
+
 const Broker =
   require(
     "../models/broker.model"
   );
+
+const AppError =
+  require(
+    "../utils/AppError"
+  );
+
+const logoDir = path.join(
+  __dirname,
+  "../../uploads/broker-logos"
+);
+
+const deleteLogoFile = (logoUrl) => {
+  if (!logoUrl) return;
+
+  const filename = logoUrl.split(
+    "/uploads/broker-logos/"
+  )[1];
+
+  if (!filename) return;
+
+  const filePath = path.join(
+    logoDir,
+    filename
+  );
+
+  if (fs.existsSync(filePath)) {
+    fs.unlinkSync(filePath);
+  }
+};
+
+const normalizeKosher = (data) => {
+  const kosherStatus =
+    data.kosherStatus ||
+    "Totally Kosher";
+
+  const kosherLine =
+    kosherStatus ===
+    "Offers Kosher Line"
+      ? (data.kosherLine || "").trim()
+      : "";
+
+  if (
+    kosherStatus ===
+      "Offers Kosher Line" &&
+    !kosherLine
+  ) {
+    throw new AppError(
+      "Kosher line is required when a broker offers a kosher line",
+      400
+    );
+  }
+
+  return { kosherStatus, kosherLine };
+};
 
 /*
 ========================================
@@ -12,9 +70,16 @@ CREATE
 exports.createBrokerService =
   async (data) => {
 
-    return await Broker.create(
-      data
-    );
+    const {
+      kosherStatus,
+      kosherLine,
+    } = normalizeKosher(data);
+
+    return await Broker.create({
+      ...data,
+      kosherStatus,
+      kosherLine,
+    });
   };
 
 /*
@@ -41,16 +106,40 @@ UPDATE
 exports.updateBrokerService =
   async (
     id,
-    data
+    data,
+    logoUrl
   ) => {
 
-    return await Broker.findByIdAndUpdate(
-      id,
-      data,
-      {
-        new: true,
-      }
-    );
+    const broker =
+      await Broker.findById(id);
+
+    if (!broker) {
+      throw new AppError(
+        "Broker not found",
+        404
+      );
+    }
+
+    const {
+      kosherStatus,
+      kosherLine,
+    } = normalizeKosher(data);
+
+    Object.assign(broker, {
+      ...data,
+      kosherStatus,
+      kosherLine,
+    });
+
+    if (logoUrl) {
+      deleteLogoFile(broker.logoUrl);
+
+      broker.logoUrl = logoUrl;
+    }
+
+    await broker.save();
+
+    return broker;
   };
 
 /*
@@ -62,7 +151,21 @@ DELETE
 exports.deleteBrokerService =
   async (id) => {
 
-    return await Broker.findByIdAndDelete(
+    const broker =
+      await Broker.findById(id);
+
+    if (!broker) {
+      throw new AppError(
+        "Broker not found",
+        404
+      );
+    }
+
+    deleteLogoFile(broker.logoUrl);
+
+    await Broker.findByIdAndDelete(
       id
     );
+
+    return true;
   };
